+++ src/controllers/EvidenceRemovalController.ts
@@ -13,8 +13,8 @@ import { FileTransferFeatureMiddleware } from 'app/middleware/FileTransferFeatur
 import { Appeal } from 'app/models/Appeal';
 import { ApplicationData, APPLICATION_DATA_KEY } from 'app/models/ApplicationData';
 import { Attachment } from 'app/models/Attachment';
-import { YesNo } from 'app/models/chunks/YesNo';
-import { createSchema } from 'app/models/chunks/YesNo.schema';
+import { YesNo } from 'app/models/fields/YesNo';
+import { createSchema } from 'app/models/fields/YesNo.schema';
 import { FileTransferService } from 'app/service/FileTransferService';
 import { EVIDENCE_REMOVAL_PAGE_URI, EVIDENCE_UPLOAD_PAGE_URI } from 'app/utils/Paths';
 import { Navigation } from 'app/utils/navigation/navigation';
