+++ src/models/fields/YesNo.schema.ts
@@ -1,6 +1,6 @@
 import Joi from '@hapi/joi';
 
-import { YesNo } from 'app/models/chunks/YesNo';
+import { YesNo } from 'app/models/fields/YesNo';
 
 export const createSchema = (errorMessage: string): Joi.StringSchema => {
     return Joi.string()
